+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ariaLabel` block support. WordPress doesn't display a UI control for this, but it's possible to change the default `Breadcrumbs` label via the Code Editor view.
 - `shadow` block support. You can now add shadows—not sure how I missed adding this before.
 - Content-only editing support for the block. When enabled, the toolbar controls no longer appear. However, users can edit the Home label directly in the editor canvas.
+- New time-based breadcrumbs:
+  - Hour, minute, and second archive breadcrumbs.
+  - Posts with hour, minute, and seconds in their permalink structure now show those crumbs (assuming rewrite mapping is enabled).
 - New `JsonLinkedData` markup class for outputting JSON-LD breadcrumbs in the site head. This is not output by default, but developers can opt in.
 - New hooks:
   - `x3p0/breadcrumbs/register` - Fires just after the plugin's default service providers are registered.